fix: Ignore Active Address for HW Verification
Now that there is support for multiple hardware ledgers with multiple
hardware accounts in the olympia wallet, there is a bug where after
significant wallet changes, the 'verify' SDK ie:
`displayAddressForActiveHWAccountOnHWDeviceForVerification` function
fires multiple requests to the `sendAPDU` tunnel with the same message.

This seems to be from the dependence on the `activeAccount` observable.
This PR removes that dependency and fixes the bug.