fix: Signing Keychain sending multiple signals (#203)
Now that the Olympia Wallet has the ability to manage multiple hardware
ledgers, the wallet consistently calls `radix.deriveHWAccount` which
requires that we pass in the `hardwareWalletConnection` observable down.
If this observable changes, (still not sure under what circumstances
this happens, but happens consistently when adding a new account on a
ledger) future actions end up sending multiple, duplicate signals
through the hardware connection.  This messaging to different addresses
seems to cause occasional race conditions and strange behavior (especially when sending
transactions with encrypted messages) about 5% of the time.