fix: Calculate Shared Secret (#204)
There's been a long standing bug in the SDK that occurs when handling
encrypted messages.  Thanks to the great debugging from the Radix
community discord, we were able to figure out whats going on and write a
simple fix to the problem.  Sometimes the shared secret calculation
returns a buffer of an unexpected length, this little fix sets the
buffers length explicitly.